util/hlc: move the PTP clock out of hlc, into timeutil
The PTP clock is not particularly tied with HLC. This patch moves it to
a dedicated package under timeutil; timeutil where most clocks live.

Release note: None
